CRUD API with middleware authenticate

+++ backend/controller/goalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModal');
+const User = require('../models/userModel')
 // @desc  Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req,res) => {
-    const goals =  await Goal.find();
+    const goals =  await Goal.find({user:req.user.id});
     res.status(200).json(goals)
 });  
 
@@ -18,7 +19,8 @@ const setGoal = asyncHandler(async (req,res) => {
     } 
     
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal);
 })
@@ -32,6 +34,18 @@ const updateGoal = asyncHandler(async (req,res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure login user matches the login user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User is not Authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
@@ -48,6 +62,20 @@ const deleteGoal = asyncHandler(async (req,res) => {
         res.status(400);
         throw new Error('Goal not found');
     }
+
+    const user = await User.findById(req.user.id)
+    
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure login user matches the login user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User is not Authorized')
+    }
+
     await goal.deleteOne();
 
     res.status(200).json({id :req.params.id})
